test(lmstudio): isolate reflected model-load transport proof
